Don't double count bridges with multiple projects
Previously, projects were counted, not unique bridges (by their bridge
keys). For example, see the below project phases which are programmed
for the Saint Charles Bridge in Armstrong County.

|BR Key|Owner                |Condition|Length (ft)|Title                  |Phase|Total  |
|------|---------------------|---------|-----------|-----------------------|-----|-------|
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |T-466 Saint Charles Br.|PE   |350000 |
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |T-466 Saint Charles Br.|FD   |350000 |
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |T-466 Saint Charles Br.|ROW  |25000  |
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |T-466 Saint Charles Br.|UTL  |20000  |
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |T-466 Saint Charles Br.|CON  |3561000|

In the past, this would have counted as five separate bridges, and
inflated the number of bridges with programmed repairs. Obviously, this
is one bridge that is being repaired across one project with five
phases.

Now, the bridge appears as follows:

|BR Key|Owner                |Condition|Length (ft)|Total                  |
|------|---------------------|---------|-----------|-----------------------|
|3378.0|COUNTY HIGHWAY AGENCY|Poor     |167.0      |4306000                |